Fix inherited suricata hook import and resolve SonarCloud findings
The dev-merge surfaced a pre-existing breakage: #543 moved
ensure_suricata_config_source_ownership from aptl.core.credentials to
aptl.core.suricata_seed but left the suricata-config-source-ownership
pre-commit hook importing the old path, which fails the all-files
pre-commit CI gate. Point the hook at the new module.

Resolve 3 SonarCloud MAJOR code smells on the DSL-008 diff:
- single-return health_status();
- extract the readiness/health-conformance cluster
  (_node_readiness_diagnostics, _warn_unhealthy_infra,
  _container_health_diagnostics, _live_container_for_node) from
  _live_gate_probes.py (over the 500-line budget) into a new
  _live_gate_readiness.py;
- _container_health_diagnostics is now single-return.
