+++ .pre-commit-config.yaml
@@ -20,7 +20,7 @@ repos:
           elif command -v python3 >/dev/null; then PY=python3;
           else echo "python3 or .venv/bin/python required" >&2; exit 1; fi;
           export PYTHONPATH="src${PYTHONPATH:+:$PYTHONPATH}";
-          exec "$PY" -c "from pathlib import Path; from aptl.core.credentials import ensure_suricata_config_source_ownership; import sys; r=ensure_suricata_config_source_ownership(Path.cwd()); (r.error and print(r.error, file=sys.stderr)); sys.exit(0 if r.success else 1)"'
+          exec "$PY" -c "from pathlib import Path; from aptl.core.suricata_seed import ensure_suricata_config_source_ownership; import sys; r=ensure_suricata_config_source_ownership(Path.cwd()); (r.error and print(r.error, file=sys.stderr)); sys.exit(0 if r.success else 1)"'
         language: system
         pass_filenames: false
         files: ^config/suricata/(suricata\.yaml|rules/local\.rules)$
